rutas follow

+++ routes/user.routes.js
@@ -1,6 +1,8 @@
 const router = require("express").Router();
 const mongoose = require("mongoose");
 const User = require("../models/User.model")
+const Booking = require("../models/Booking.model")
+const Review = require("../models/Review.model")
 const verifyToken = require("../middlewares/auth.middlewares");
 
 
@@ -8,7 +10,18 @@ const verifyToken = require("../middlewares/auth.middlewares");
 router.get("/", async (req, res, next)=>{
     try {
         const response = await User.find()
-        res.json(response)
+        res.json(response.data)
+    } catch (error) {
+        console.log(error);
+        next(error);
+    }
+})
+
+// GET "/api/users" => Ver todos los usuarios
+router.get("/userby/:userId", async (req, res, next)=>{
+    try {
+        const response = await User.findById(req.params.userId)
+        res.status(200).json(response)
     } catch (error) {
         console.log(error);
         next(error);
@@ -63,12 +76,20 @@ router.patch("/profile", verifyToken, async (req, res, next)=>{
 // DELETE "/api/users/profile" => Eliminar un usuario concreto
 // router.delete("/:userId", async(req, res, next)=>{
 router.delete("/profile", verifyToken, async(req, res, next)=>{
-    try {
-        const response = await User.findByIdAndDelete(req.payload._id)
-        res.json({message:`El usuario ${response.name} ha sido eliminado.`})
-    } catch (error) {
-        console.log(error);
-        next(error)
+    const userId = req.payload._id
+    if (userId){
+        try {
+            await Review.deleteMany({user: userId});
+            await Booking.deleteMany({user: userId});
+            const response = await User.findByIdAndDelete(req.payload._id)
+            res.json({message:`El usuario ${response.name} ha sido eliminado.`})
+        } catch (error) {
+            console.log(error);
+            next(error)
+        }
+    }
+    else{
+        res.status(404).json({message: "No estás autorizado a hacer eso"})
     }
 })
 
@@ -120,7 +141,22 @@ router.put("/unfav/:restaurantId", verifyToken, async (req, res, next)=>{
             {$pull: {wishlist: restaurantId}}
         )
         res.status(200).json({message: `Restaurante eliminado de tu lista de deseos`})
-    } catch (error) {
+    }
+    catch (error) {
+        console.log(error)
+        next(error)
+    }
+})
+
+// GET "/api/users/allfavs/:restaurantId"
+router.get("/allfavs/:restaurantId", async (req, res, next)=>{
+    const {restaurantId} = req.params
+    // console.log(restaurantId)
+    try {
+        const response = await User.find({wishlist: restaurantId})
+        res.status(200).json(response.length)
+    }
+    catch (error) {
         console.log(error)
         next(error)
     }
@@ -145,6 +181,72 @@ router.get("/owner", verifyToken, async (req, res, next)=>{
     } catch (error) {
         console.log(error);
         next(error);
+    }
+})
+
+// PUT "/api/users/follow/:userId"
+router.put("/follow/:userId", verifyToken, async (req, res, next)=>{
+    try {
+        const {userId} = req.params
+        const loggedUser = req.payload._id
+        const usuario = await User.findById(userId)
+        const response = await User.findByIdAndUpdate(
+            loggedUser,
+            {$addToSet: {follow: userId}}
+        )
+        res.status(200).json({message: `Ahora sigues a ${usuario.username}`})
+    } catch (error) {
+        console.log(error)
+        next(error)
+    }
+})
+
+// PUT "/api/users/unfollow/:userId"
+router.put("/unfollow/:userId", verifyToken, async (req, res, next)=>{
+    try {
+        const {userId} = req.params
+        const loggedUser = req.payload._id
+        const usuario = await User.findById(userId)
+        const response = await User.findByIdAndUpdate(
+            loggedUser,
+            {$pull: {follow: userId}}
+        )
+        res.status(200).json({message: `Has dejado de seguir a ${usuario.username}`})
+    } catch (error) {
+        console.log(error)
+        next(error)
+    }
+})
+
+// GET "/api/users/is-following/:userId"
+router.get("/is-following/:userId", verifyToken, async (req, res, next)=>{
+    try {
+        const {userId} = req.params;
+        const loggedUser = req.payload._id
+        const response = await User.findById(loggedUser, 'follow')
+        // console.log(response)
+        if (response.follow.includes(userId)){
+            res.status(200).json(true)
+        }
+        else{
+            res.status(200).json(false)
+        }
+        
+    } catch (error) {
+        next(error)
+    }
+})
+
+// GET "/api/users/followers"
+router.get("/followers", verifyToken, async (req, res, next)=>{
+    try {
+        const loggedUser = req.payload._id
+        const response = await User.find({follow:loggedUser})
+        // console.log(response.length)
+        res.status(200).json(response.length)
+        
+    } catch (error) {
+        next(error)
     }
 })
